Update README.md
This is a Flipkart Clone built using HTML and CSS,. The project mimics the basic design and layout of the popular e-commerce website Flipkart, with responsive features, interactive UI elements, and a user-friendly design. The aim of this project is to demonstrate front-end development skills by creating a functional and visually appealing web page.